+++ packages/react-hook/src/useEffectOnce/index.ts
@@ -3,10 +3,12 @@ import { useEffect, useRef } from 'react'
 function useEffectOnce(cb: React.EffectCallback, deps?: React.DependencyList | undefined): void {
   const mountRef = useRef(false)
   useEffect(() => {
+    let returnValue
     if (mountRef.current === false) {
-      cb()
+      returnValue = cb()
       mountRef.current = true
     }
+    return returnValue
   }, [deps])
 }