+++ packages/vite-config/src/node/plugins/public-typescript.ts
@@ -1,4 +1,4 @@
-import { type VitePublicTypescriptOptions, publicTypescript } from 'vite-plugin-public-typescript'
+import { publicTypescript, type VitePublicTypescriptOptions } from 'vite-plugin-public-typescript'
 
 export function pt(options: VitePublicTypescriptOptions) {
   return publicTypescript({