+++ packages/vite-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @minko-fe/vite-config
 
+## 1.1.0
+
+### Minor Changes
+
+- fix use client warn
+
 ## 1.0.2
 
 ### Patch Changes